Update README.md
One of the issues that occur during neural network training is called Overfitting. A model can become excessively complex in order to capture all the relations inherently expressed by the training data. The network has memorized the training examples, but it is not able to generalize to new situations.
In order to solve the overfitting problem, I have applied Regularization techniques like Weight regularization and dropout regularizarion. Regularization can be a good way to increase the performance of a network, particularly when there is an evident situation of overfitting.
Here, i have applied Regularization on the popular fashion_mnist dataset.